pyproject.toml
[project]
name = "autonotes"
version = "0.1.0"
description = "AI Orchestrator for Obsidian vault management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "celery[redis]>=5.4.0",
    "httpx>=0.28.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "python-frontmatter>=1.1.0",
    "ruamel.yaml>=0.18.0,<0.19",
    "markdown-it-py>=3.0.0,<4",
    "mdit-py-plugins>=0.4.0,<0.5",
    "redis>=5.2.0",
    "anthropic>=0.42.0",
    "openai>=1.60.0",
    "pgvector>=0.4.2",
    "numpy>=2.4.3",
    "scikit-learn>=1.8.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
]
